refactor popup
